+++ config/dev.exs
@@ -1,4 +1,1 @@
 import Config
-
-config :opentelemetry,
-  processors: [{:otel_batch_processor, %{scheduled_delay_ms: 1}}]