+++ lib/open_telemetry_decorator/attributes.ex
@@ -1,9 +1,22 @@
 defmodule OpenTelemetryDecorator.Attributes do
   @moduledoc false
   require OpenTelemetry.Tracer, as: Tracer
+  require OpenTelemetry.Span, as: Span
 
-  def set(name, value) do
-    Tracer.set_attribute(name, to_otlp_value(value))
+  # This is getting way too fucking complicated
+
+  def set(span_ctx, name, value) do
+    Span.set_attribute(span_ctx, Atom.to_string(prefix_name(name)), to_otlp_value(value))
+  end
+
+  def set(name, value) when is_binary(name) or is_atom(name) do
+    set(Tracer.current_span_ctx(), name, value)
+  end
+
+  def set(span_ctx, attributes) do
+    Enum.map(attributes, fn {key, value} ->
+      set(span_ctx, key, value)
+    end)
   end
 
   def set(attributes) when is_struct(attributes) do
@@ -11,9 +24,7 @@ defmodule OpenTelemetryDecorator.Attributes do
   end
 
   def set(attributes) do
-    attributes
-    |> Enum.map(fn {key, value} -> {key, to_otlp_value(value)} end)
-    |> Tracer.set_attributes()
+    set(Tracer.current_span_ctx(), attributes)
   end
 
   def get(all_attributes, requested_attributes) do
@@ -67,6 +78,9 @@ defmodule OpenTelemetryDecorator.Attributes do
     joiner = Application.get_env(:open_telemetry_decorator, :attr_joiner) || "."
     Enum.join(keys, joiner)
   end
+
+  defp prefix_name(:error), do: :error
+  defp prefix_name("error"), do: :error
 
   defp prefix_name(name) when is_atom(name), do: prefix_name(Atom.to_string(name))